+++ core/src/test/java/cucumber/runtime/RuntimeOptionsFactoryTest.java
@@ -1,8 +1,7 @@
-package cucumber.runtime.junit;
+package cucumber.runtime;
 
+import cucumber.api.CucumberOptions;
 import cucumber.api.SnippetType;
-import cucumber.api.junit.Cucumber;
-import cucumber.runtime.RuntimeOptions;
 import gherkin.formatter.Formatter;
 import gherkin.formatter.JSONFormatter;
 import gherkin.formatter.PrettyFormatter;
@@ -14,8 +13,9 @@ import java.util.Iterator;
 import java.util.List;
 import java.util.regex.Pattern;
 
-import static cucumber.runtime.junit.RuntimeOptionsFactory.packageName;
-import static cucumber.runtime.junit.RuntimeOptionsFactory.packagePath;
+import static cucumber.runtime.RuntimeOptionsFactory.packageName;
+import static cucumber.runtime.RuntimeOptionsFactory.packagePath;
+import static java.util.Arrays.asList;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
@@ -23,35 +23,37 @@ import static org.junit.Assert.assertTrue;
 public class RuntimeOptionsFactoryTest {
     @Test
     public void create_strict() throws Exception {
-        RuntimeOptionsFactory runtimeOptionsFactory = new RuntimeOptionsFactory(Strict.class);
+        RuntimeOptionsFactory runtimeOptionsFactory = new RuntimeOptionsFactory(Strict.class, new Class[]{CucumberOptions.class});
         RuntimeOptions runtimeOptions = runtimeOptionsFactory.create();
         assertTrue(runtimeOptions.isStrict());
     }
 
     @Test
     public void create_non_strict() throws Exception {
-        RuntimeOptionsFactory runtimeOptionsFactory = new RuntimeOptionsFactory(NotStrict.class);
+        RuntimeOptionsFactory runtimeOptionsFactory = new RuntimeOptionsFactory(NotStrict.class, new Class[]{CucumberOptions.class});
         RuntimeOptions runtimeOptions = runtimeOptionsFactory.create();
         assertFalse(runtimeOptions.isStrict());
     }
 
     @Test
     public void create_without_options() throws Exception {
-        RuntimeOptionsFactory runtimeOptionsFactory = new RuntimeOptionsFactory(WithoutOptions.class);
+        RuntimeOptionsFactory runtimeOptionsFactory = new RuntimeOptionsFactory(WithoutOptions.class, new Class[]{CucumberOptions.class});
         RuntimeOptions runtimeOptions = runtimeOptionsFactory.create();
         assertFalse(runtimeOptions.isStrict());
+        assertEquals(asList("classpath:cucumber/runtime"), runtimeOptions.getFeaturePaths());
+        assertEquals(asList("classpath:cucumber/runtime"), runtimeOptions.getGlue());
     }
 
     @Test
     public void create_with_no_name() throws Exception {
-        RuntimeOptionsFactory runtimeOptionsFactory = new RuntimeOptionsFactory(NoName.class);
+        RuntimeOptionsFactory runtimeOptionsFactory = new RuntimeOptionsFactory(NoName.class, new Class[]{CucumberOptions.class});
         RuntimeOptions runtimeOptions = runtimeOptionsFactory.create();
         assertTrue(runtimeOptions.getFilters().isEmpty());
     }
 
     @Test
     public void create_with_multiple_names() throws Exception {
-        RuntimeOptionsFactory runtimeOptionsFactory = new RuntimeOptionsFactory(MultipleNames.class);
+        RuntimeOptionsFactory runtimeOptionsFactory = new RuntimeOptionsFactory(MultipleNames.class, new Class[]{CucumberOptions.class});
 
         RuntimeOptions runtimeOptions = runtimeOptionsFactory.create();
 
@@ -64,21 +66,21 @@ public class RuntimeOptionsFactoryTest {
 
     @Test
     public void create_with_dotcucumber_dir() throws MalformedURLException {
-        RuntimeOptionsFactory runtimeOptionsFactory = new RuntimeOptionsFactory(DotCucumberFile.class);
+        RuntimeOptionsFactory runtimeOptionsFactory = new RuntimeOptionsFactory(DotCucumberFile.class, new Class[]{CucumberOptions.class});
         RuntimeOptions runtimeOptions = runtimeOptionsFactory.create();
         assertEquals(new URL("file:somewhere/.cucumber/"), runtimeOptions.getDotCucumber());
     }
 
     @Test
     public void create_with_dotcucumber_url() throws MalformedURLException {
-        RuntimeOptionsFactory runtimeOptionsFactory = new RuntimeOptionsFactory(DotCucumberUrl.class);
+        RuntimeOptionsFactory runtimeOptionsFactory = new RuntimeOptionsFactory(DotCucumberUrl.class, new Class[]{CucumberOptions.class});
         RuntimeOptions runtimeOptions = runtimeOptionsFactory.create();
         assertEquals(new URL("https://some.where/.cucumber/"), runtimeOptions.getDotCucumber());
     }
 
     @Test
     public void create_with_snippets() {
-        RuntimeOptionsFactory runtimeOptionsFactory = new RuntimeOptionsFactory(Snippets.class);
+        RuntimeOptionsFactory runtimeOptionsFactory = new RuntimeOptionsFactory(Snippets.class, new Class[]{CucumberOptions.class});
         RuntimeOptions runtimeOptions = runtimeOptionsFactory.create();
         assertEquals(SnippetType.CAMELCASE, runtimeOptions.getSnippetType());
     }
@@ -99,7 +101,7 @@ public class RuntimeOptionsFactoryTest {
 
     @Test
     public void inherit_formatter_from_baseclass() {
-        RuntimeOptionsFactory runtimeOptionsFactory = new RuntimeOptionsFactory(SubClassWithFormatter.class);
+        RuntimeOptionsFactory runtimeOptionsFactory = new RuntimeOptionsFactory(SubClassWithFormatter.class, new Class[]{CucumberOptions.class});
         RuntimeOptions runtimeOptions = runtimeOptionsFactory.create();
 
         List<Formatter> formatters = runtimeOptions.getFormatters();
@@ -110,44 +112,44 @@ public class RuntimeOptionsFactoryTest {
 
     @Test
     public void override_monochrome_flag_from_baseclass() {
-        RuntimeOptionsFactory runtimeOptionsFactory = new RuntimeOptionsFactory(SubClassWithMonoChromeTrue.class);
+        RuntimeOptionsFactory runtimeOptionsFactory = new RuntimeOptionsFactory(SubClassWithMonoChromeTrue.class, new Class[]{CucumberOptions.class});
         RuntimeOptions runtimeOptions = runtimeOptionsFactory.create();
 
         assertTrue(runtimeOptions.isMonochrome());
     }
 
 
-    @Cucumber.Options(snippets = SnippetType.CAMELCASE)
+    @CucumberOptions(snippets = SnippetType.CAMELCASE)
     static class Snippets {
         // empty
     }
 
-    @Cucumber.Options(strict = true)
+    @CucumberOptions(strict = true)
     static class Strict {
         // empty
     }
 
-    @Cucumber.Options
+    @CucumberOptions
     static class NotStrict {
         // empty
     }
 
-    @Cucumber.Options(name = {"name1", "name2"})
+    @CucumberOptions(name = {"name1", "name2"})
     static class MultipleNames {
         // empty
     }
 
-    @Cucumber.Options
+    @CucumberOptions
     static class NoName {
         // empty
     }
 
-    @Cucumber.Options(dotcucumber = "somewhere/.cucumber")
+    @CucumberOptions(dotcucumber = "somewhere/.cucumber")
     static class DotCucumberFile {
         // empty
     }
 
-    @Cucumber.Options(dotcucumber = "https://some.where/.cucumber")
+    @CucumberOptions(dotcucumber = "https://some.where/.cucumber")
     static class DotCucumberUrl {
         // empty
     }
@@ -156,22 +158,22 @@ public class RuntimeOptionsFactoryTest {
         // empty
     }
 
-    @Cucumber.Options(format = "pretty")
+    @CucumberOptions(format = "pretty")
     static class SubClassWithFormatter extends BaseClassWithFormatter {
         // empty
     }
 
-    @Cucumber.Options(format = "json:outFile")
+    @CucumberOptions(format = "json:test-json-report.json")
     static class BaseClassWithFormatter {
         // empty
     }
 
-    @Cucumber.Options(monochrome = true)
+    @CucumberOptions(monochrome = true)
     static class SubClassWithMonoChromeTrue extends BaseClassWithMonoChromeFalse {
         // empty
     }
 
-    @Cucumber.Options(monochrome = false)
+    @CucumberOptions(monochrome = false)
     static class BaseClassWithMonoChromeFalse {
         // empty
     }