+++ examples/java-calculator/src/test/java/cucumber/examples/java/calculator/RunCukesTest.java
@@ -1,9 +1,10 @@
 package cucumber.examples.java.calculator;
 
+import cucumber.api.CucumberOptions;
 import cucumber.api.junit.Cucumber;
 import org.junit.runner.RunWith;
 
 @RunWith(Cucumber.class)
-@Cucumber.Options(format = "json:target/cucumber-report.json")
+@CucumberOptions(format = "json:target/cucumber-report.json")
 public class RunCukesTest {
 }