+++ junit/src/test/java/cucumber/runtime/junit/CucumberTest.java
@@ -11,7 +11,9 @@ import org.junit.runners.model.InitializationError;
 
 import java.io.File;
 import java.io.IOException;
+import java.util.List;
 
+import static java.util.Collections.emptyList;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.fail;
 
@@ -35,7 +37,7 @@ public class CucumberTest {
 
     @Test
     public void finds_features_based_on_implicit_package() throws IOException, InitializationError {
-        Cucumber cucumber = new Cucumber(ImplicitFeaturePath.class);
+        Cucumber cucumber = new Cucumber(ImplicitFeatureAndGluePath.class);
         assertEquals(3, cucumber.getChildren().size());
         assertEquals("Feature: FA", cucumber.getChildren().get(0).getName());
     }
@@ -58,9 +60,10 @@ public class CucumberTest {
     }
 
     @Test
-    public void finds_no_features_when_explicit_package_has_nothnig() throws IOException, InitializationError {
+    public void finds_no_features_when_explicit_feature_path_has_no_features() throws IOException, InitializationError {
         Cucumber cucumber = new Cucumber(ExplicitFeaturePathWithNoFeatures.class);
-        assertEquals(0, cucumber.getChildren().size());
+        List<FeatureRunner> children = cucumber.getChildren();
+        assertEquals(emptyList(), children);
     }
 
     @RunWith(Cucumber.class)
@@ -91,7 +94,7 @@ public class CucumberTest {
         Assertions.assertNoCucumberAnnotatedMethods(RunCukesTestInvalid.class);
     }
 
-    private class ImplicitFeaturePath {
+    private class ImplicitFeatureAndGluePath {
     }
 
     @Cucumber.Options(features = {"classpath:cucumber/runtime/junit"})